requirements.txt
fastapi~=0.115.12
sqlalchemy~=2.0.41
uvicorn
pytest~=8.3.5
httpx
pydantic~=2.11.5
starlette~=0.46.2
python-dotenv
pytest-mock